@@ -9,8 +9,9 @@
     "test-server": "NODE_ENV=test JASMINE_CONFIG_PATH=test/server/jasmine.json babel-istanbul cover --dir test/server/coverage jasmine",
     "test": "npm run test-client",
     "test-ci": "NODE_ENV=test karma start ./test/client/karma.travis.conf.js",
-    "build": "rimraf lib && NODE_ENV=production gulp",
-    "precommit": "npm run lint && npm test"
+    "build:dev": "NODE_ENV=production rollup --config tools/rollup.dev.js",
+    "build:min": "NODE_ENV=production rollup --config tools/rollup.min.js",
+    "build": "rimraf lib && npm run build:dev && npm run build:min"
   },
   "author": {
     "name": "ludafa@outlook.com"
@@ -18,36 +19,29 @@
   "repository": "https://github.com/jinzhubaofu/ei",
   "license": "ISC",
   "peerDependencies": {
-    "react": "^0.14.0||^15.4.0"
+    "react": "^16.0.0"
   },
   "readme": "readme.md",
   "devDependencies": {
-    "babel-istanbul": "^0.11.0",
     "babel-loader": "^6.2.5",
-    "babel-plugin-external-helpers": "^6.8.0",
-    "babel-plugin-external-helpers-insert-require": "^1.0.2",
+    "babel-plugin-external-helpers": "^6.22.0",
     "babel-plugin-istanbul": "^3.0.0",
-    "babel-plugin-transform-es2015-modules-umd": "^6.12.0",
     "babel-plugin-transform-es3-member-expression-literals": "^6.8.0",
     "babel-plugin-transform-es3-property-literals": "^6.8.0",
-    "babel-plugin-transform-object-rest-spread": "^6.8.0",
     "babel-preset-es2015": "^6.14.0",
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-2": "^6.13.0",
-    "babel-register": "^6.14.0",
-    "babelify": "^7.3.0",
     "coveralls": "^2.11.12",
-    "enzyme": "^2.6.0",
+    "enzyme": "^3.1.0",
+    "enzyme-adapter-react-16": "^1.0.2",
     "esdoc": "^0.4.8",
     "esdoc-es7-plugin": "0.0.3",
     "fecs": "^0.8.7",
     "gulp": "^3.9.1",
     "gulp-babel": "^6.1.2",
-    "gulp-babel-external-helpers": "^2.0.2",
     "gulp-clean": "^0.3.2",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-webpack": "^1.5.0",
-    "istanbul-lib-coverage": "^1.0.0",
     "jasmine": "^2.4.1",
     "json-loader": "^0.5.4",
     "karma": "^1.2.0",
@@ -60,56 +54,19 @@
     "karma-webpack": "^1.8.0",
     "lcov-result-merger": "^1.2.0",
     "pre-commit": "^1.1.2",
-    "react": "^0.14.0||^15.4.0",
-    "react-addons-test-utils": "^0.14.0||^15.4.0",
-    "react-dom": "^0.14.0||^15.4.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "rimraf": "^2.5.4",
+    "rollup": "^0.50.0",
+    "rollup-plugin-babel": "^3.0.2",
+    "rollup-plugin-babel-minify": "^3.1.2",
     "vinyl-named": "^1.1.0",
     "watchify": "^3.7.0",
-    "webpack": "^1.13.2",
-    "webpack-babel-external-helpers": "^0.1.0",
-    "webpack-stream": "^3.2.0"
-  },
-  "babel": {
-    "env": {
-      "dev": {
-        "presets": [
-          "es2015",
-          "stage-2",
-          "react"
-        ]
-      },
-      "test": {
-        "presets": [
-          "es2015",
-          "stage-2",
-          "react"
-        ],
-        "plugins": [
-          "istanbul"
-        ]
-      },
-      "production": {
-        "presets": [
-          [
-            "es2015",
-            {
-              "loose": true,
-              "modules": "umd"
-            }
-          ],
-          "stage-2",
-          "react"
-        ],
-        "plugins": [
-          "transform-es3-property-literals",
-          "transform-es3-member-expression-literals"
-        ]
-      }
-    }
+    "webpack": "^1.13.2"
   },
   "dependencies": {
+    "prop-types": "^15.6.0",
     "react-redux": "^4.4.6",
-    "redux": "^3.6.0",
-    "rimraf": "^2.5.4"
+    "redux": "^3.6.0"
   }
 }
