@@ -3,6 +3,6 @@
  * @author leon(ludafa@outlook.com)
  */
 
-module.exports = function () {
+export default function () {
     return Math.random().toString(36).substr(2, 12);
 };
