@@ -3,24 +3,20 @@
  * @author leon(ludafa@outlook.com)
  */
 
-const INIT = 'ei/INIT';
+export const INIT = 'ei/INIT';
 
-exports.INIT = INIT;
-
-exports.init = function (payload) {
+export function init(payload) {
     return {
         type: INIT,
         payload
     };
-};
+}
 
-const REPLACE = 'ei/REPLACE';
+export const REPLACE = 'ei/REPLACE';
 
-exports.REPLACE = REPLACE;
-
-exports.replace = function (payload) {
+export function replace(payload) {
     return {
         type: REPLACE,
         payload
     };
-};
+}
